refactor: moved function to a separate file and fix ts errors
